+++ android/src/main/kotlin/dev/note11/flutter_naver_map/flutter_naver_map/view/NaverMapViewFactory.kt
@@ -29,6 +29,7 @@ internal class NaverMapViewFactory(
 
         return NaverMapView(
             activity = activity,
+            flutterProvidedContext = context,
             naverMapViewOptions = options,
             channel = channel,
             overlayController = overlayController,