+++ example/lib/main.dart
@@ -1,12 +1,17 @@
 import 'dart:async';
+import 'dart:math';
 
 import 'package:flutter/material.dart';
 import 'package:flutter_naver_map/flutter_naver_map.dart';
+import 'package:flutter_naver_map_example/pages/examples/camera_example.dart';
 import 'package:flutter_naver_map_example/pages/examples/controller_example.dart';
 import 'package:flutter_naver_map_example/pages/examples/overlay_example.dart';
+import 'package:flutter_naver_map_example/pages/examples/pick_example.dart';
+import 'package:flutter_naver_map_example/util/overlay_portal_util.dart';
+import 'package:transparent_pointer/transparent_pointer.dart';
 
-import 'pages/bottom_drawer.dart';
-import 'pages/new_window_page.dart';
+import 'pages/utils/bottom_drawer.dart';
+import 'pages/utils/new_window_page.dart';
 import 'design/map_function_item.dart';
 import 'design/theme.dart';
 
@@ -49,16 +54,26 @@ class _FNMapPageState extends State<FNMapPage> {
   late EdgeInsets safeArea;
   double drawerHeight = 0;
 
+  final nOverlayInfoOverlayPortalController = NInfoOverlayPortalController();
+
   @override
   Widget build(BuildContext context) {
     safeArea = MediaQuery.of(context).padding;
-    return WillPopScope(
-      onWillPop: () async => drawerTool.processWillPop(),
-      child: Stack(children: [
-        Positioned.fill(child: mapWidget()),
-        drawerTool.bottomDrawer,
-      ]),
-    );
+    return Scaffold(
+        body: PopScope(
+            onPopInvoked: (didPop) => drawerTool.processWillPop(),
+            child: Stack(children: [
+              GestureDetector(
+                  onTapDown: (details) => _onLastTouchStreamController.sink
+                      .add(details.globalPosition)),
+              Positioned.fill(child: TransparentPointer(child: mapWidget())),
+              drawerTool.bottomDrawer,
+              OverlayPortal(
+                  controller: nOverlayInfoOverlayPortalController,
+                  overlayChildBuilder: (context) =>
+                      nOverlayInfoOverlayPortalController.builder(
+                          context, mapController)),
+            ])));
   }
 
   /*
@@ -87,7 +102,7 @@ class _FNMapPageState extends State<FNMapPage> {
     mapController = controller;
   }
 
-  void onMapTapped(NPoint point, NLatLng latLng) {
+  void onMapTapped(NPoint point, NLatLng latLng) async {
     // ...
   }
 
@@ -108,7 +123,8 @@ class _FNMapPageState extends State<FNMapPage> {
     // ...
   }
 
-  final _onCameraChangeStreamController = StreamController.broadcast();
+  final _onCameraChangeStreamController = StreamController<void>.broadcast();
+  final _onLastTouchStreamController = StreamController<Offset>.broadcast();
 
   /*
     --- Bottom Drawer Widget ---
@@ -122,8 +138,9 @@ class _FNMapPageState extends State<FNMapPage> {
 
   late final List<MapFunctionItem> pages = [
     MapFunctionItem(
-        title: "NaverMapViewOptions 변경",
-        description: "지도의 옵션을 변경할 수 있어요",
+        title: "지도 위젯 옵션 변경하기",
+        description: "위젯에 보여지는 걸 바꿔봐요",
+        icon: Icons.map_rounded,
         page: (canScroll) => NaverMapViewOptionsExample(
             canScroll: canScroll,
             options: options,
@@ -132,61 +149,54 @@ class _FNMapPageState extends State<FNMapPage> {
             })),
     MapFunctionItem(
         title: "오버레이 추가 / 제거",
-        description: "마커, 경로 등의 각종 오버레이들을 추가하고 제거할 수 있어요",
+        description: "마커/경로/도형 등을 띄워봐요",
+        icon: Icons.add_location_alt_rounded,
         isScrollPage: false,
         page: (canScroll) => NOverlayExample(
-            canScroll: canScroll, mapController: mapController)),
+            nOverlayInfoOverlayPortalController:
+                nOverlayInfoOverlayPortalController,
+            canScroll: canScroll,
+            mapController: mapController)),
     MapFunctionItem(
         title: "카메라 이동",
-        description: "지도에 보이는 영역을 카메라를 이동하여 바꿀 수 있어요",
-        page: (canScroll) => _cameraMoveTestPage(mapController)),
+        isScrollPage: false,
+        icon: Icons.zoom_in_rounded,
+        description: "지도를 요리조리 움직여봐요",
+        page: (canScroll) => CameraUpdateExample(
+            onCameraChangeStream: _onCameraChangeStreamController.stream,
+            canScroll: canScroll,
+            mapController: mapController)),
+    MapFunctionItem(
+        title: "주변 Pickable 보기",
+        description: "주변 심볼, 오버레이를 찾아봐요",
+        icon: Icons.domain_rounded,
+        page: (canScroll) {
+          final screenSize = MediaQuery.sizeOf(context);
+          return NaverMapPickExample(
+            canScroll: canScroll,
+            mapController: mapController,
+            mapEndPoint:
+                Point(screenSize.width, screenSize.height - drawerHeight),
+            onCameraChangeStream: _onCameraChangeStreamController.stream,
+          );
+        }),
     MapFunctionItem(
         title: "기타 컨트롤러 기능",
-        description: "컨트롤러로 지도의 상태를 가져오거나 변경할 수 있습니다.",
+        description: "컨트롤러 기능을 살펴봐요",
         isScrollPage: false,
+        icon: Icons.sports_esports_rounded,
         page: (canScroll) => NaverMapControllerExample(
               canScroll: canScroll,
               mapController: mapController,
               onCameraChangeStream: _onCameraChangeStreamController.stream,
+              onLastTouchStream: _onLastTouchStreamController.stream,
             )),
     MapFunctionItem(
-        title: "주변 심볼 및 오버레이 가져오기",
-        description: "특정 영역 주변의 심볼 및 오버레이를 가져올 수 있어요",
-        page: (canScroll) => _pickTestPage()),
-    MapFunctionItem(
       title: "새 페이지에서 지도 보기",
-      description: "새 페이지에서 지도를 봅니다. (메모리 누수 확인용)",
+      icon: Icons.note_add_rounded,
+      description: "테스트용이에요",
       onTap: (_) => Navigator.of(context).push(
           MaterialPageRoute(builder: (context) => const NewWindowTestPage())),
     ),
   ];
-
-  Widget _cameraMoveTestPage(NaverMapController mapController) {
-    return Padding(
-        padding: const EdgeInsets.all(24),
-        child: Column(children: [
-          // todo
-          const Text("_cameraMoveTestPage"),
-          const Text("카메라 이동"),
-          ElevatedButton(
-              onPressed: () {
-                mapController.updateCamera(NCameraUpdate.fromCameraPosition(
-                    const NCameraPosition(
-                        target: NLatLng(37.56362422812855, 126.96269803941277),
-                        zoom: 17.00922642853924,
-                        bearing: 119.62995870263971)));
-              },
-              child: const Text('카메라 회전')),
-        ]));
-  }
-
-  Widget _pickTestPage() {
-    return const Padding(
-        padding: EdgeInsets.all(24),
-        child: Column(children: [
-          // todo
-          Text("_pickTestPage"),
-          Text("주변 심볼 및 오버레이 가져오기"),
-        ]));
-  }
 }