+++ lib/src/type/default/padding.dart
@@ -1,8 +1,7 @@
 part of flutter_naver_map;
 
 class NEdgeInsets extends EdgeInsets with NMessageableWithMap {
-  const NEdgeInsets.fromLTRB(
-      super.left, super.top, super.right, super.bottom)
+  const NEdgeInsets.fromLTRB(super.left, super.top, super.right, super.bottom)
       : super.fromLTRB();
 
   NEdgeInsets.fromEdgeInsets(EdgeInsets insets)